+++ eras/conway/impl/test/Main.hs
@@ -12,6 +12,7 @@ import qualified Test.Cardano.Ledger.Conway.GoldenTranslation as GoldenTranslati
 import qualified Test.Cardano.Ledger.Conway.GovActionReorderSpec as GovActionReorder
 import Test.Cardano.Ledger.Conway.Plutus.PlutusSpec as PlutusSpec
 import qualified Test.Cardano.Ledger.Conway.Spec as ConwaySpec
+import Test.Cardano.Ledger.Shelley.JSON (roundTripJsonShelleyEraSpec)
 
 main :: IO ()
 main = ledgerTestMain $ do
@@ -31,3 +32,4 @@ main = ledgerTestMain $ do
       let step = 25600
       map (tierRefScriptFee 1.5 step 15) [0, step .. 204800]
         `shouldBe` map Coin [0, 384000, 960000, 1824000, 3120000, 5064000, 7980000, 12354000, 18915000]
+  roundTripJsonShelleyEraSpec @ConwayEra